+++ src/plugins/schema-check.js
@@ -238,7 +238,8 @@ export function checkSchema(jsonID) {
         .reduce((a, b) => a.concat(b), [])
         .filter((elem, pos, arr) => arr.indexOf(elem) === pos) // unique
         .map(key => {
-            const schemaObj = objectPath.get(jsonID, 'properties.' + key.replace(/\./g, '.properties.'));
+            const path = 'properties.' + key.replace(/\./g, '.properties.');
+            const schemaObj = objectPath.get(jsonID, path);
             if (!schemaObj || typeof schemaObj !== 'object') {
                 throw RxError.newRxError('SC21', {
                     key