+++ src/util.js
@@ -329,3 +329,28 @@ export const clone = recursiveDeepCopy;
 
 import isElectron from 'is-electron';
 export const isElectronRenderer = isElectron();
+
+
+/**
+ * returns a flattened object
+ * @link https://gist.github.com/penguinboy/762197
+ */
+export function flattenObject(ob) {
+    const toReturn = {};
+
+    for (const i in ob) {
+        if (!ob.hasOwnProperty(i)) continue;
+
+        if ((typeof ob[i]) == 'object') {
+            const flatObject = flattenObject(ob[i]);
+            for (const x in flatObject) {
+                if (!flatObject.hasOwnProperty(x)) continue;
+
+                toReturn[i + '.' + x] = flatObject[x];
+            }
+        } else {
+            toReturn[i] = ob[i];
+        }
+    }
+    return toReturn;
+};